requirements.txt
pip
autopep8
pandas
numpy
accelerate
peft
bitsandbytes
transformers
trl
llama-index
llama-index-embeddings-openai
llama-index-embeddings-huggingface
orjson==3.9.15 
langchain==0.1.9
langchain-community==0.0.21
faiss-gpu
nest_asyncio
openai
python-dotenv
llama-index-vector-stores-faiss
llama-index-retrievers-bm25
rank_bm25

